+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v0.2.0
+
+### Added
+
+- `--offline` flag for `pint lint` command. When passed only checks that don't send
+  any live queries to Prometheus server will be run.
+- `alerts/template` check will now warn if template if referencing a label that is being
+  stripped by aggregation.
+  Example:
+
+  ```
+  - alert: Foo
+    expr: count(up) without(instance) == 0
+    annotations:
+      summary: 'foo is down on {{ $labels.instance }}'
+  ```
+
+  Would generate a warning since `instance` label is being stripped by `without(instance)`.
+
 ## v0.1.5
 
 ### Fixed